refactored tests to separate files
