+++ crates/astria-composer/src/searcher/executor/bundle_factory/tests.rs
@@ -1,177 +1,3 @@
-use std::{
-    collections::VecDeque,
-    mem,
-};
-
-use astria_core::sequencer::v1alpha1::{
-    transaction::{
-        action::SequenceAction,
-        Action,
-    },
-    ROLLUP_ID_LEN,
-};
-use tracing::trace;
-
-#[derive(Debug, thiserror::Error)]
-enum SizedBundleError {
-    #[error("bundle does not have enough space left for the given sequence action")]
-    NotEnoughSpace(SequenceAction),
-    #[error("sequence action is larger than the max bundle size")]
-    SequenceActionTooLarge(SequenceAction),
-}
-
-/// A bundle sequence actions to be submitted to the sequencer. Maintains the total size of the
-/// bytes pushed to it and enforces a max size in bytes passed in the constructor. If an incoming
-/// `seq_action` won't fit in the buffer it is flushed and a new bundle is started.
-struct SizedBundle {
-    /// The buffer of actions
-    buffer: Vec<Action>,
-    /// The current size of the bundle in bytes. This is equal to the sum of the size of the
-    /// `seq_action`s + `ROLLUP_ID_LEN` for each.
-    curr_size: usize,
-    /// The max bundle size in bytes to enforce.
-    max_size: usize,
-}
-
-impl SizedBundle {
-    /// Create a new empty bundle with the given max size.
-    fn new(max_size: usize) -> Self {
-        Self {
-            buffer: vec![],
-            curr_size: 0,
-            max_size,
-        }
-    }
-
-    /// Buffer `seq_action` into the bundle. If the bundle won't fit `seq_action`, flush `buffer`,
-    /// returning it, and start building up a new buffer using `seq_action`.
-    fn push(&mut self, seq_action: SequenceAction) -> Result<(), SizedBundleError> {
-        let seq_action_size = estimate_size_of_sequence_action(&seq_action);
-
-        if seq_action_size > self.max_size {
-            return Err(SizedBundleError::SequenceActionTooLarge(seq_action));
-        }
-
-        if self.curr_size + seq_action_size > self.max_size {
-            return Err(SizedBundleError::NotEnoughSpace(seq_action));
-        }
-
-        self.buffer.push(Action::Sequence(seq_action));
-        self.curr_size += seq_action_size;
-        Ok(())
-    }
-
-    /// Replace self with a new empty bundle, returning the old bundle.
-    fn flush(&mut self) -> SizedBundle {
-        mem::replace(self, Self::new(self.max_size))
-    }
-
-    /// Consume self and return the underlying buffer of actions.
-    fn into_actions(self) -> Vec<Action> {
-        self.buffer
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub(super) enum BundleFactoryError {
-    #[error("sequence action is larger than the max bundle size. seq_action size: {size}")]
-    SequenceActionTooLarge { size: usize, max_size: usize },
-}
-
-/// Manages the bundling of sequence actions into `SizedBundle`s. A `Vec<Action>` is flushed and
-/// added to the `finished` queue when an incoming `SequenceAction` won't fit in the current bundle.
-/// The `finished` queue operates in FIFO order, where `Vec<Action>`s are added to the back and
-/// taken off from the front.
-pub(super) struct BundleFactory {
-    /// The current bundle being built.
-    curr_bundle: SizedBundle,
-    /// The queue of bundles that have been built but not yet sent to the sequencer.
-    finished: VecDeque<SizedBundle>,
-}
-
-impl BundleFactory {
-    pub(super) fn new(max_bytes_per_bundle: usize) -> Self {
-        Self {
-            curr_bundle: SizedBundle::new(max_bytes_per_bundle),
-            finished: VecDeque::new(),
-        }
-    }
-
-    /// Buffer `seq_action` into the current bundle. If the bundle won't fit `seq_action`, flush
-    /// `curr_bundle` into the `finished` queue and start a new bundle
-    pub(super) fn try_push(
-        &mut self,
-        seq_action: SequenceAction,
-    ) -> Result<(), BundleFactoryError> {
-        let seq_action_size = estimate_size_of_sequence_action(&seq_action);
-
-        match self.curr_bundle.push(seq_action) {
-            Err(SizedBundleError::SequenceActionTooLarge(_seq_action)) => {
-                // reject the sequence action if it is larger than the max bundle size
-                Err(BundleFactoryError::SequenceActionTooLarge {
-                    size: seq_action_size,
-                    max_size: self.curr_bundle.max_size,
-                })
-            }
-            Err(SizedBundleError::NotEnoughSpace(seq_action)) => {
-                // if the bundle is full, flush it and start a new one
-                self.finished.push_back(self.curr_bundle.flush());
-                self.curr_bundle
-                    .push(seq_action)
-                    .expect("seq_action should not be larger than max bundle size");
-                Ok(())
-            }
-            Ok(()) => {
-                trace!(
-                    new_bundle_size = self.curr_bundle.curr_size,
-                    seq_action_size = seq_action_size,
-                    "bundled new sequence action"
-                );
-                Ok(())
-            }
-        }
-    }
-
-    /// Get the next bundle from the `finished` queue.
-    pub(super) fn next_finished(&mut self) -> Option<NextFinishedBundle> {
-        if self.finished.is_empty() {
-            None
-        } else {
-            Some(NextFinishedBundle {
-                bundle_factory: self,
-            })
-        }
-    }
-
-    /// Get the next bundle from the `finished` queue. If the queue is empty, flush `curr_bundle`.
-    pub(super) fn pop_now(&mut self) -> Vec<Action> {
-        self.finished
-            .pop_front()
-            .or_else(|| Some(self.curr_bundle.flush()))
-            .map(SizedBundle::into_actions)
-            .unwrap_or_default()
-    }
-}
-
-pub(super) struct NextFinishedBundle<'a> {
-    bundle_factory: &'a mut BundleFactory,
-}
-
-impl<'a> NextFinishedBundle<'a> {
-    pub(super) fn pop(self) -> Vec<Action> {
-        self.bundle_factory
-            .finished
-            .pop_front()
-            .map(SizedBundle::into_actions)
-            .expect("next bundle exists. this is a bug.")
-    }
-}
-
-/// The size of the `seq_action` in bytes, including the rollup id.
-fn estimate_size_of_sequence_action(seq_action: &SequenceAction) -> usize {
-    seq_action.data.len() + ROLLUP_ID_LEN
-}
-
 #[cfg(test)]
 mod sized_bundle_tests {
     use astria_core::sequencer::v1alpha1::{
@@ -180,19 +6,19 @@ mod sized_bundle_tests {
         ROLLUP_ID_LEN,
     };
 
-    use super::SizedBundle;
-    use crate::searcher::bundle_factory::{
+    use crate::searcher::executor::bundle_factory::{
         estimate_size_of_sequence_action,
+        SizedBundle,
         SizedBundleError,
     };
 
     #[test]
     fn push_works() {
         // create a bundle with 100 bytes of max space
-        let mut bundle = super::SizedBundle::new(100);
+        let mut bundle = SizedBundle::new(100);
 
         // push a sequence action that is 100 bytes total
-        let seq_action = super::SequenceAction {
+        let seq_action = SequenceAction {
             rollup_id: RollupId::new([0; ROLLUP_ID_LEN]),
             data: vec![0; 100 - ROLLUP_ID_LEN],
         };
@@ -204,10 +30,10 @@ mod sized_bundle_tests {
     #[test]
     fn push_seq_action_too_large() {
         // create a bundle with 100 bytes of max space
-        let mut bundle = super::SizedBundle::new(100);
+        let mut bundle = SizedBundle::new(100);
 
         // push a sequence action that is >100 bytes total
-        let seq_action = super::SequenceAction {
+        let seq_action = SequenceAction {
             rollup_id: RollupId::new([0; ROLLUP_ID_LEN]),
             data: vec![0; 100 - ROLLUP_ID_LEN + 1],
         };
@@ -281,8 +107,11 @@ mod bundle_factory_tests {
         ROLLUP_ID_LEN,
     };
 
-    use super::BundleFactory;
-    use crate::searcher::bundle_factory::estimate_size_of_sequence_action;
+    use crate::searcher::executor::bundle_factory::{
+        estimate_size_of_sequence_action,
+        BundleFactory,
+        BundleFactoryError,
+    };
 
     #[test]
     fn try_push_works_no_flush() {
@@ -314,7 +143,7 @@ mod bundle_factory_tests {
 
         assert!(matches!(
             bundle_factory.try_push(seq_action),
-            Err(super::BundleFactoryError::SequenceActionTooLarge {
+            Err(BundleFactoryError::SequenceActionTooLarge {
                 size,
                 max_size
             }) if size == actual_size && max_size == 100
